tidy: Fix "multiplication implicit widening" in libmythtv (2).
Fix the clang-tidy "implicit widening of multiplication result"
warnings.  This check is looking for places where a multiplication
operation may have overflowed the smaller type before the result is
converted to the bigger type. This will give erroneous results.

1) Cast multiplication arguments to a larger type.

2) Convert variables to the appropriate type.

3) Convert a couple of #define to typed static constexpr.

+++ mythtv/libs/libmythtv/recorders/HLS/HLSReader.h
@@ -45,7 +45,7 @@ class MTV_PUBLIC  HLSReader
 
     bool Open(const QString & m3u, int bitrate_index = 0);
     void Close(bool quiet = false);
-    int Read(uint8_t* buffer, int len);
+    qint64 Read(uint8_t* buffer, qint64 len);
     void Throttle(bool val);
     bool IsThrottled(void) const { return m_throttle; }
     bool IsOpen(const QString& url) const
